+++ pos/imports/ui/pages/receiveItem/receiveItem.js
@@ -119,11 +119,38 @@ indexTmpl.events({
     },
     'click .js-destroy' (event, instance) {
         let data = this;
-        destroyAction(
-            ReceiveItems,
-            {_id: data._id},
-            {title: TAPi18n.__('pos.receiveItem.title'), itemTitle: data._id}
-        );
+        let inventoryDate = InventoryDates.findOne({branchId: data.branchId, stockLocationId: data.stockLocationId});
+        let receiveItemDate = moment(data.receiveItemDate).startOf('days').toDate();
+        if (inventoryDate && (receiveItemDate < inventoryDate.inventoryDate)) {
+            swal({
+                title: "Date is less then current Transaction Date!",
+                text: "Stock will recalculate on: '" + moment(inventoryDate.inventoryDate).format("DD-MM-YYYY") + "'",
+                type: "warning", showCancelButton: true,
+                confirmButtonColor: "#DD6B55",
+                confirmButtonText: "Yes, Do it!",
+                closeOnConfirm: false
+            }).then(function () {
+                swal.close();
+                destroyAction(
+                    ReceiveItems,
+                    {_id: data._id},
+                    {title: TAPi18n.__('pos.receiveItem.title'), itemTitle: data._id}
+                );
+            }, function (dismiss) {
+                if (dismiss === 'cancel') {
+                    return false;
+                }
+            });
+        }
+        else {
+            destroyAction(
+                ReceiveItems,
+                {_id: data._id},
+                {title: TAPi18n.__('pos.receiveItem.title'), itemTitle: data._id}
+            );
+        }
+
+
     },
     'click .js-display' (event, instance) {
         swal({
